Move `std::io` trait impls to their own modules in `spinoso-string`
Each string type has an impl of `std::io::Write`. All such impls have
been moved to an `io` module local to each type:

- `crate::enc::ascii::io`
- `crate::enc::binary::io`
- `crate::enc::utf8::io`
- `crate::enc::io`

This helps with code symmetry, makes the imports easier to read, moves
some logic out of the monolithic `impls` modules, and limits `io::Write`
code to a single `#[cfg(feature = "std")]` attribute.

Fixes #1724.
